Split the message processing to a separate file
